+++ server/src/main/java/com/unosquare/carmigo/controller/AuthenticationController.java
@@ -4,6 +4,7 @@ import com.unosquare.carmigo.dto.CreateAuthenticationDTO;
 import com.unosquare.carmigo.dto.GrabAuthenticationDTO;
 import com.unosquare.carmigo.model.request.CreateAuthenticationViewModel;
 import com.unosquare.carmigo.model.response.AuthenticationViewModel;
+import com.unosquare.carmigo.service.AuthenticationService;
 import com.unosquare.carmigo.service.PlatformUserService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import javax.validation.Valid;
@@ -25,7 +26,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthenticationController {
 
   private final ModelMapper modelMapper;
-  private final PlatformUserService platformUserService;
+  private final AuthenticationService authenticationService;
 
   @PostMapping(value = "/login", produces = MediaType.APPLICATION_JSON_VALUE)
   @ResponseStatus(HttpStatus.CREATED)
@@ -34,7 +35,7 @@ public class AuthenticationController {
     final CreateAuthenticationDTO createAuthenticationDTO = modelMapper.map(
         createAuthenticationViewModel, CreateAuthenticationDTO.class);
     final GrabAuthenticationDTO grabAuthenticationDTO =
-        platformUserService.createAuthenticationToken(createAuthenticationDTO);
+        authenticationService.createAuthenticationToken(createAuthenticationDTO);
     final AuthenticationViewModel authenticationViewModel = modelMapper.map(
         grabAuthenticationDTO, AuthenticationViewModel.class);
     return new ResponseEntity<>(authenticationViewModel, HttpStatus.CREATED);
