+++ server/src/test/java/com/unosquare/carmigo/controller/AuthenticationControllerTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.unosquare.carmigo.service.PlatformUserService;
+import com.unosquare.carmigo.service.AuthenticationService;
 import com.unosquare.carmigo.util.ControllerUtility;
 import com.unosquare.carmigo.util.ResourceUtility;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,14 +29,14 @@ public class AuthenticationControllerTest {
   private MockMvc mockMvc;
 
   @Mock private ModelMapper modelMapperMock;
-  @Mock private PlatformUserService platformUserServiceMock;
+  @Mock private AuthenticationService authenticationServiceMock;
 
   private ControllerUtility controllerUtility;
 
   @BeforeEach
   public void setUp() {
     mockMvc = MockMvcBuilders.standaloneSetup(
-            new AuthenticationController(modelMapperMock, platformUserServiceMock))
+            new AuthenticationController(modelMapperMock, authenticationServiceMock))
         .build();
 
     controllerUtility = new ControllerUtility(mockMvc, API_LEADING);
@@ -45,12 +45,12 @@ public class AuthenticationControllerTest {
   @Test
   public void post_Create_Authentication_Token_Returns_HttpStatus_Created() throws Exception {
     controllerUtility.makePostRequest("login", POST_AUTHENTICATION_VALID_JSON, status().isCreated());
-    verify(platformUserServiceMock).createAuthenticationToken(any());
+    verify(authenticationServiceMock).createAuthenticationToken(any());
   }
 
   @Test
   public void post_Create_Authentication_Token_Returns_HttpStatus_BadRequest() throws Exception {
     controllerUtility.makePostRequest("login", POST_AUTHENTICATION_INVALID_JSON, status().isBadRequest());
-    verify(platformUserServiceMock, times(0)).createAuthenticationToken(any());
+    verify(authenticationServiceMock, times(0)).createAuthenticationToken(any());
   }
 }
